serverless/requirements.txt
psycopg[binary]==3.1.2
python-dotenv==0.21.0
